Moved schedule preview list items to their own component
* This makes it so we can useRef() to dynamically place buttons
