+++ autoscheduler/frontend/src/components/SchedulingPage/SchedulePreview/ScheduleListItem/MiniSchedule/MiniSchedule.css
